requirements.txt
faster-coco-eval>=1.6.5
PyYAML
tensorboard
scipy
calflops
thop
transformers
pytorch_wavelets==1.3.0
timm==1.0.7
grad-cam==1.5.4
tidecv
einops
prettytable
pycocotools==2.0.8

# ------------------ 装不成功可以先注释这两个
mmengine==0.10.7
mmcv==2.2.0

# engine/extre_module/custom_nn/neck_module/HS_FPN.py 需要的
torch-dct==0.1.6